+++ data-raw/DATASET.R
@@ -1,9 +1,10 @@
 ## code to prepare `DATASET` dataset goes here
-
+library(evavelo)
 
 ## Dataset example for tests
 xlsx_path <- system.file("example-data/02_simplified.xlsx", package = "evavelo")
 evavelo_example <- read_evavelo(xlsx_path)
+evavelo_example_geocoded <- geocode_evavelo(evavelo_example)
 
 ## id_quest_mismatch
 mismatch_path <- system.file("example-data/03_id_quest_mismatch.xlsx", package = "evavelo")
@@ -26,6 +27,7 @@ enquete_colnames <- c(
   "iti_km_voyage", "iti_depart_initial", "iti_depart_itineraire", "iti_arrivee_itineraire",
   "iti_arrivee_final", "iti_experience",
   "distance_domicile_enq", "distance_heb_enq","distance_dom_enq_reelle", "distance_heb_enq_reelle",
+  "mode_transp_jour", "dist_transp_jour",
   "iti_dep_iti_valide", "iti_arr_iti_valide",
   "ville_heb_cog_lau", "ville_res_cog_lau",
   "id_section_origine", "id_section_dest", "taille_totale_groupe")
@@ -33,6 +35,7 @@ enquete_colnames <- c(
 calendrier_colnames <- c("id_site_enq", "date_enq")
 
 
-usethis::use_data(evavelo_example, quest_mismatch_example, all_enquete_example,
+usethis::use_data(evavelo_example, evavelo_example_geocoded,
+                  quest_mismatch_example, all_enquete_example,
                   comptage_colnames, enquete_colnames, calendrier_colnames,
                   overwrite = TRUE, internal = TRUE)