+++ R/app_ui.R
@@ -10,7 +10,6 @@ app_ui <- function(request) {
     #addResourcePath('img', system.file('app/img', package = 'evavelo')),
     # List the first level UI elements here
     fluidPage(
-      #theme = bslib::bs_theme(version = 4, bg = "black", fg = "white"),
       theme = bslib::bs_theme(bootswatch = "yeti"),
       shinyjs::useShinyjs(),
       fluidPage(fluidRow(
@@ -23,9 +22,16 @@ app_ui <- function(request) {
       )),
       titlePanel("EVA-Scan : outil de traitement des donn\u00e9es collect\u00e9es selon la m\u00e9thode \u00c9VA-V\u00c9LO"),
       em("Application d\u00e9velopp\u00e9e dans le cadre du projet Interreg AtlanticOnBike et cofinanc\u00e9 par l\u2019Union Europ\u00e9enne"),
+      br(),br(),
       sidebarLayout(
         sidebarPanel(
           mod_load_file_ui("load_file_ui_1"),
+          h3("Scan"),
+          mod_check_file_ui("check_file_ui_1"),
+          br(), br(),
+          h3("Process"),
+          mod_process_file_ui("process_file_ui_1"),
+          br(),br(),
           mod_download_result_ui("download_result_ui_1"),
           br(),br(),
           p("D\u00e9veloppements r\u00e9alis\u00e9s sur la base de la ",