+++ NEWS.md
@@ -1,4 +1,4 @@
-# evavelo (development version)
+# evavelo 0.6.0
 
 - Add find_wrong_geocoding() to export all the cities that couldn't be geocoded.